complete project seven chapter seven

+++ cModernApproach/chapter7projects/three.c
@@ -2,20 +2,20 @@
 
 int main(int argc, char const *argv[])
 {
-    float n, sum = 0;
+    double n, sum = 0;
 
 
     printf("This program sums a series of double.\n");
     printf("Enter integers (0 to terminate): ");
 
-    scanf("%f", &n);
+    scanf("%lf", &n);
     while (n != 0)
     {
         sum += n;
-        scanf("%f", &n);
+        scanf("%lf", &n);
     }
 
-    printf("\nThe sum is: %f\n", sum);
+    printf("\nThe sum is: %lf\n", sum);
     
     return 0;
 }
